Fix Fedora 32 gcc/clang warnings
Component: AOSP, OpenGL, Vulkan

VK-GL-CTS issue: 2350

Affects:
KHR*direct_state_access.vertex_arrays*
KHR*core.arrays_of_arrays.sized_declarations_invalid*
dEQP-VK.geometry.*

(cherry picked from commit a4ccc4a84710123839a4f4a9409e93af9ed1ddbc)

+++ external/vulkancts/modules/vulkan/clipping/vktClippingTests.cpp
@@ -386,7 +386,7 @@ void initPrograms (SourceCollections& programCollection, const LineOrientation l
 
 void initProgramsPointSize (SourceCollections& programCollection)
 {
-	addSimplePrograms(programCollection, 0.75f * RENDER_SIZE);
+	addSimplePrograms(programCollection, 0.75f * static_cast<float>(RENDER_SIZE));
 }
 
 //! Primitives fully inside the clip volume.
