+++ external/vulkancts/modules/vulkan/pipeline/vktPipelineBlendTests.cpp
@@ -585,8 +585,8 @@ float BlendTestInstance::getNormChannelThreshold (const tcu::TextureFormat& form
 {
 	switch (tcu::getTextureChannelClass(format.type))
 	{
-		case tcu::TEXTURECHANNELCLASS_UNSIGNED_FIXED_POINT:	return BlendTest::QUAD_COUNT / static_cast<float>((1 << numBits) - 1);
-		case tcu::TEXTURECHANNELCLASS_SIGNED_FIXED_POINT:	return BlendTest::QUAD_COUNT / static_cast<float>((1 << (numBits - 1)) - 1);
+		case tcu::TEXTURECHANNELCLASS_UNSIGNED_FIXED_POINT:	return static_cast<float>(BlendTest::QUAD_COUNT) / static_cast<float>((1 << numBits) - 1);
+		case tcu::TEXTURECHANNELCLASS_SIGNED_FIXED_POINT:	return static_cast<float>(BlendTest::QUAD_COUNT) / static_cast<float>((1 << (numBits - 1)) - 1);
 		default:
 			break;
 	}
